# stockwellness-api/src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Stockwellness API
  description: 자산 배분 시뮬레이터 및 AI 예측 서비스 API 명세서
  version: 0.0.1
servers:
- url: http://localhost:8080
...
paths:
...
  /api/v1/portfolios/{portfolioId}/analysis/backtest:
    post:
      tags:
      - Portfolio Analysis
      summary: 포트폴리오 과거 시뮬레이션(백테스팅) 실행
      description: 포트폴리오 과거 시뮬레이션(백테스팅) 실행
      operationId: portfolio-analysis-backtest
      parameters:
      - name: portfolioId
        in: path
        description: 포트폴리오 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api-v1-portfolios-portfolioId-analysis-backtest-617935080"
            examples:
              portfolio-analysis-backtest:
                value: |-
                  {
                    "strategy" : "LUMP_SUM",
                    "amount" : 1000000,
                    "benchmarkTicker" : "KOSPI",
                    "period" : "ALL",
                    "rebalancingPeriod" : "MONTHLY",
                    "weights" : {
                      "005930" : 100
                    }
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-portfolios-portfolioId-analysis-backtest-1617317571"
              examples:
                portfolio-analysis-backtest:
                  value: |-
                    {
                      "success" : true,
                      "status" : 200,
                      "code" : "S000",
                      "message" : "요청이 성공적으로 처리되었습니다.",
                      "data" : {
                        "dailyResults" : [ {
                          "date" : "2024-01-01",
                          "totalValue" : 1000000,
                          "totalInvested" : 1000000,
                          "returnRate" : 5.0,
                          "benchmarkReturnRate" : 3.0,
                          "benchmarkReturnRates" : {
                            "KOSPI" : 3.0
                          }
                        } ],
                        "cagr" : 0.15,
                        "mdd" : -0.1,
                        "relativeMdd" : -0.05,
                        "sharpeRatio" : 1.5,
                        "totalReturnRate" : 0.2,
                        "volatility" : 0.12,
                        "alpha" : 0.05,
                        "beta" : 1.0,
                        "bestYearRate" : 0.25,
                        "worstYearRate" : -0.05,
                        "itemReturns" : {
                          "005930" : 10.0
                        },
                        "comparisons" : [ {
                          "indexName" : "코스피",
                          "ticker" : "KOSPI",
                          "totalReturn" : 15.5,
                          "alpha" : 4.5,
                          "beta" : 1.0,
                          "mdd" : 10.0,
                          "relativeMdd" : -5.0
                        } ],
                        "aiComment" : "현재 포트폴리오는 시장 지수 대비 안정적인 수익을 보여주고 있습니다."
                      },
                      "timestamp" : "2026-04-07T15:50:36.402728",
                      "traceId" : null,
                      "errors" : [ ]
                    }
  /api/v1/portfolios/{portfolioId}/analysis/correlation:
    get:
      tags:
      - Portfolio Analysis
      summary: 포트폴리오 종목 간 상관관계 분석
      description: 포트폴리오 종목 간 상관관계 분석
      operationId: portfolio-analysis-correlation
      parameters:
      - name: portfolioId
        in: path
        description: 포트폴리오 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-portfolios-portfolioId-analysis-correlation-116670467"
              examples:
                portfolio-analysis-correlation:
                  value: |-
                    {
                      "success" : true,
                      "status" : 200,
                      "code" : "S000",
                      "message" : "요청이 성공적으로 처리되었습니다.",
                      "data" : {
                        "AAPL" : {
                          "AAPL" : 1,
                          "TSLA" : 0.85
                        },
                        "TSLA" : {
                          "AAPL" : 0.85,
                          "TSLA" : 1
                        }
                      },
                      "timestamp" : "2026-04-07T15:50:36.340379",
                      "traceId" : null,
                      "errors" : [ ]
                    }
  /api/v1/portfolios/{portfolioId}/analysis/diversification:
    get:
      tags:
      - Portfolio Analysis
      summary: 포트폴리오 자산 분산도 분석
      description: 포트폴리오 자산 분산도 분석
      operationId: portfolio-analysis-diversification
      parameters:
      - name: portfolioId
        in: path
        description: 포트폴리오 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-portfolios-portfolioId-analysis-diversification-1682980467"
              examples:
                portfolio-analysis-diversification:
                  value: |-
                    {
                      "success" : true,
                      "status" : 200,
                      "code" : "S000",
                      "message" : "요청이 성공적으로 처리되었습니다.",
                      "data" : {
                        "totalValue" : 1500000,
                        "assetRatios" : [ {
                          "name" : "STOCK",
                          "value" : 60.0
                        } ],
                        "sectorRatios" : [ {
                          "name" : "TECH",
                          "value" : 40.0
                        } ],
                        "countryRatios" : [ {
                          "name" : "US",
                          "value" : 100.0
                        } ]
                      },
                      "timestamp" : "2026-04-07T15:50:36.361811",
                      "traceId" : null,
                      "errors" : [ ]
                    }
  /api/v1/portfolios/{portfolioId}/analysis/rebalancing:
    get:
      tags:
      - Portfolio Analysis
      summary: 포트폴리오 리밸런싱 분석
      description: 포트폴리오 리밸런싱 분석
      operationId: portfolio-analysis-rebalancing
      parameters:
      - name: portfolioId
        in: path
        description: 포트폴리오 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-portfolios-portfolioId-analysis-rebalancing1494527279"
              examples:
                portfolio-analysis-rebalancing:
                  value: |-
                    {
                      "success" : true,
                      "status" : 200,
                      "code" : "S000",
                      "message" : "요청이 성공적으로 처리되었습니다.",
                      "data" : {
                        "totalValue" : 1500000,
                        "items" : [ {
                          "symbol" : "AAPL",
                          "currentWeight" : 50.0,
                          "targetWeight" : 60.0,
                          "diffWeight" : -10.0,
                          "currentQuantity" : 10.0,
                          "recommendedQuantity" : 12.0,
                          "currentPrice" : 150000,
                          "expectedTradeAmount" : 1800000
                        } ]
                      },
                      "timestamp" : "2026-04-07T15:50:36.473793",
                      "traceId" : null,
                      "errors" : [ ]
                    }
  /api/v1/portfolios/{portfolioId}/analysis/summary:
    get:
      tags:
      - Portfolio Analysis
      summary: 포트폴리오 통합 분석 요약
      description: 포트폴리오 통합 분석 요약
      operationId: portfolio-analysis-summary
      parameters:
      - name: portfolioId
        in: path
        description: 포트폴리오 ID
        required: true
        schema:
          type: string
      - name: startDate
        in: query
        description: 분석 시작일 (YYYY-MM-DD)
        required: false
        schema:
          type: string
      - name: endDate
        in: query
        description: 분석 종료일 (YYYY-MM-DD)
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-portfolios-portfolioId-analysis-summary1301469393"
              examples:
                portfolio-analysis-summary:
                  value: |-
                    {
                      "success" : true,
                      "status" : 200,
                      "code" : "S000",
                      "message" : "요청이 성공적으로 처리되었습니다.",
                      "data" : {
                        "valuation" : {
                          "totalPurchaseAmount" : 1350000,
                          "currentTotalValue" : 1500000,
                          "totalProfitLoss" : 150000,
                          "totalReturnRate" : 12.5,
                          "dailyProfitLoss" : 0,
                          "dailyReturnRate" : 0.0,
                          "cagr" : 15.5,
                          "volatility" : 12.0,
                          "alpha" : 3.5,
                          "mdd" : 10.2,
                          "sharpeRatio" : 1.5,
                          "beta" : 1.1,
                          "totalInstitutionalNetBuying" : 0,
                          "totalForeignNetBuying" : 0
                        },
                        "diversification" : {
                          "totalValue" : 1500000,
                          "assetRatios" : [ {
                            "name" : "STOCK",
                            "value" : 100.0
                          } ],
                          "sectorRatios" : [ {
                            "name" : "TECH",
                            "value" : 100.0
                          } ],
                          "countryRatios" : [ {
                            "name" : "US",
                            "value" : 100.0
                          } ]
                        },
                        "rebalancing" : {
                          "totalValue" : 1500000,
                          "items" : [ ]
                        },
                        "itemContributions" : {
                          "AAPL" : 12.5
                        }
                      },
                      "timestamp" : "2026-04-07T15:50:36.249495",
                      "traceId" : null,
                      "errors" : [ ]
                    }
  /api/v1/portfolios/{portfolioId}/analysis/valuation:
    get:
      tags:
      - Portfolio Analysis
      summary: 포트폴리오 자산 가치 분석
      description: 포트폴리오 자산 가치 분석
      operationId: portfolio-analysis-value
      parameters:
      - name: portfolioId
        in: path
        description: 포트폴리오 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-v1-portfolios-portfolioId-analysis-valuation1906810676"
              examples:
                portfolio-analysis-value:
                  value: |-
                    {
                      "success" : true,
                      "status" : 200,
                      "code" : "S000",
                      "message" : "요청이 성공적으로 처리되었습니다.",
                      "data" : {
                        "totalPurchaseAmount" : 1350000,
                        "currentTotalValue" : 1500000,
                        "totalProfitLoss" : 150000,
                        "totalReturnRate" : 12.5,
                        "dailyProfitLoss" : 0,
                        "dailyReturnRate" : 0.0,
                        "cagr" : 15.5,
                        "volatility" : 12.0,
                        "alpha" : 3.5,
                        "mdd" : 10.2,
                        "sharpeRatio" : 1.5,
                        "beta" : 1.1,
                        "totalInstitutionalNetBuying" : 0,
                        "totalForeignNetBuying" : 0
                      },
                      "timestamp" : "2026-04-07T15:50:36.497363",
                      "traceId" : null,
                      "errors" : [ ]
                    }
...
components:
  schemas:
...
    api-v1-portfolios-portfolioId-analysis-summary1301469393:
      required:
      - code
      - message
      - status
      - success
      - timestamp
      type: object
      properties:
        code:
          type: string
          description: 비즈니스 상세 코드
        data:
          required:
          - diversification
          - itemContributions
          - rebalancing
          - valuation
          type: object
          properties:
            valuation:
              type: object
              description: 가치 평가 정보
            itemContributions:
              type: object
              description: 종목별 수익 기여도 (Ticker -> Contribution)
            diversification:
              type: object
              description: 분산도 정보
            rebalancing:
              type: object
              description: 리밸런싱 정보
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 결과 메시지
        errors:
          type: array
          description: 상세 필드 에러 목록 (성공 시 빈 리스트)
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        timestamp:
          type: string
          description: 응답 시간
        status:
          type: number
          description: HTTP 상태 코드
...
    api-v1-portfolios-portfolioId-analysis-valuation1906810676:
      required:
      - code
      - message
      - status
      - success
      - timestamp
      type: object
      properties:
        code:
          type: string
          description: 비즈니스 상세 코드
        data:
          required:
          - alpha
          - beta
          - cagr
          - currentTotalValue
          - dailyProfitLoss
          - dailyReturnRate
          - mdd
          - sharpeRatio
          - totalForeignNetBuying
          - totalInstitutionalNetBuying
          - totalProfitLoss
          - totalPurchaseAmount
          - totalReturnRate
          - volatility
          type: object
          properties:
            dailyReturnRate:
              type: number
              description: 당일 수익률 (%)
            dailyProfitLoss:
              type: number
              description: 당일 평가 손익
            cagr:
              type: number
              description: 연평균 성장률 (CAGR)
            totalInstitutionalNetBuying:
              type: number
              description: 총 기관 순매수 금액
            volatility:
              type: number
              description: 연간 변동성
            mdd:
              type: number
              description: 최대 낙폭 (MDD)
            totalReturnRate:
              type: number
              description: 총 수익률 (%)
            totalProfitLoss:
              type: number
              description: 총 평가 손익
            alpha:
              type: number
              description: 초과 수익률 (Alpha)
            sharpeRatio:
              type: number
              description: 샤프 지수
            totalPurchaseAmount:
              type: number
              description: 총 매수 금액
            currentTotalValue:
              type: number
              description: 현재 총 자산 가치
            beta:
              type: number
              description: 베타 계수
            totalForeignNetBuying:
              type: number
              description: 총 외국인 순매수 금액
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 결과 메시지
        errors:
          type: array
          description: 상세 필드 에러 목록 (성공 시 빈 리스트)
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        timestamp:
          type: string
          description: 응답 시간
        status:
          type: number
          description: HTTP 상태 코드
...
    api-v1-portfolios-portfolioId-analysis-diversification-1682980467:
      required:
      - code
      - message
      - status
      - success
      - timestamp
      type: object
      properties:
        code:
          type: string
          description: 비즈니스 상세 코드
        data:
          required:
          - totalValue
          type: object
          properties:
            totalValue:
              type: number
              description: 총 자산 가치
            countryRatios:
              type: array
              items:
                required:
                - name
                - value
                type: object
                properties:
                  name:
                    type: string
                    description: 국가 명칭
                  value:
                    type: number
                    description: 비중 (%)
            sectorRatios:
              type: array
              items:
                required:
                - name
                - value
                type: object
                properties:
                  name:
                    type: string
                    description: 섹터 명칭
                  value:
                    type: number
                    description: 비중 (%)
            assetRatios:
              type: array
              items:
                required:
                - name
                - value
                type: object
                properties:
                  name:
                    type: string
                    description: 자산군 명칭
                  value:
                    type: number
                    description: 비중 (%)
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 결과 메시지
        errors:
          type: array
          description: 상세 필드 에러 목록 (성공 시 빈 리스트)
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        timestamp:
          type: string
          description: 응답 시간
        status:
          type: number
          description: HTTP 상태 코드
...
    api-v1-portfolios-portfolioId-analysis-backtest-1617317571:
      required:
      - code
      - message
      - status
      - success
      - timestamp
      type: object
      properties:
        code:
          type: string
          description: 비즈니스 상세 코드
        data:
          required:
          - alpha
          - bestYearRate
          - beta
          - cagr
          - itemReturns
          - mdd
          - relativeMdd
          - sharpeRatio
          - totalReturnRate
          - volatility
          - worstYearRate
          type: object
          properties:
            dailyResults:
              type: array
              items:
                required:
                - benchmarkReturnRate
                - benchmarkReturnRates
                - date
                - returnRate
                - totalInvested
                - totalValue
                type: object
                properties:
                  totalValue:
                    type: number
                    description: 해당 일자 총 자산 가치
                  date:
                    type: string
                    description: 시뮬레이션 일자
                  benchmarkReturnRate:
                    type: number
                    description: 주요 벤치마크 누적 수익률 (%) — 요청한 benchmarkTicker 기준
                  returnRate:
                    type: number
                    description: 해당 일자 누적 수익률 (%)
                  totalInvested:
                    type: number
                    description: 해당 일자 총 누적 투자금
                  benchmarkReturnRates:
                    type: object
                    description: "벤치마크 지수별 해당 일자 수익률 (Map<Ticker, Rate>)"
            cagr:
              type: number
              description: 연평균 복리 수익률 (CAGR)
            volatility:
              type: number
              description: 수익률 표준편차 (변동성)
            bestYearRate:
              type: number
              description: 최고 수익을 기록한 해의 수익률
            mdd:
              type: number
              description: 최대 낙폭 (MDD)
            relativeMdd:
              type: number
              description: 벤치마크 대비 상대 낙폭
            worstYearRate:
              type: number
              description: 최저 수익을 기록한 해의 수익률
            aiComment:
              type: string
              description: AI 엔진이 생성한 백테스트 분석 코멘트
              nullable: true
            totalReturnRate:
              type: number
              description: 전체 기간 총 수익률
            itemReturns:
              type: object
              description: 종목별 수익률 기여도
            alpha:
              type: number
              description: 벤치마크 대비 초과 수익률 (Alpha)
            comparisons:
              type: array
              items:
                required:
                - alpha
                - beta
                - indexName
                - mdd
                - relativeMdd
                - ticker
                - totalReturn
                type: object
                properties:
                  ticker:
                    type: string
                    description: 비교 지수 티커
                  totalReturn:
                    type: number
                    description: 비교 지수의 전체 기간 총 수익률
                  indexName:
                    type: string
                    description: "비교 지수 명칭 (예: 코스피, 나스닥)"
                  alpha:
                    type: number
                    description: 지수 대비 해당 포트폴리오의 초과 수익
                  mdd:
                    type: number
                    description: 비교 지수의 MDD
                  relativeMdd:
                    type: number
                    description: 지수 대비 해당 포트폴리오의 상대 낙폭
                  beta:
                    type: number
                    description: 지수 대비 해당 포트폴리오의 베타
            sharpeRatio:
              type: number
              description: 위험 대비 수익 지수 (샤프 지수)
            beta:
              type: number
              description: 시장 지수 변동성 대비 민감도 (Beta)
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 결과 메시지
        errors:
          type: array
          description: 상세 필드 에러 목록 (성공 시 빈 리스트)
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        timestamp:
          type: string
          description: 응답 시간
        status:
          type: number
          description: HTTP 상태 코드
...
    api-v1-portfolios-portfolioId-analysis-backtest-617935080:
      required:
      - amount
      - benchmarkTicker
      - strategy
      type: object
      properties:
        period:
          type: string
          description: 클라이언트 기간 필터링 호환용 필드 (서버 계산에는 미반영)
          nullable: true
        amount:
          type: number
          description: 초기 투자 금액 (또는 월간 적립액)
        weights:
          type: object
          description: 사용자 정의 종목별 비중 (미입력 시 현재 포트폴리오 비중 유지)
          nullable: true
        strategy:
          type: string
          description: "투자 전략 (LUMP_SUM: 거액 적립, DCA: 정기 적립)"
        rebalancingPeriod:
          type: string
          description: "리밸런싱 주기 (NONE, MONTHLY, QUARTERLY, YEARLY)"
          nullable: true
        benchmarkTicker:
          type: string
          description: "비교 대상 대표 벤치마크 티커 (예: KOSPI, SPY)"
...
    api-v1-portfolios-portfolioId-analysis-correlation-116670467:
      required:
      - code
      - data
      - message
      - status
      - success
      - timestamp
      type: object
      properties:
        code:
          type: string
          description: 비즈니스 상세 코드
        data:
          type: object
          description: 종목별 상관관계 행렬 데이터 (가변 구조)
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 결과 메시지
        errors:
          type: array
          description: 상세 필드 에러 목록 (성공 시 빈 리스트)
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        timestamp:
          type: string
          description: 응답 시간
        status:
          type: number
          description: HTTP 상태 코드
...
    api-v1-portfolios-portfolioId-analysis-rebalancing1494527279:
      required:
      - code
      - message
      - status
      - success
      - timestamp
      type: object
      properties:
        code:
          type: string
          description: 비즈니스 상세 코드
        data:
          required:
          - totalValue
          type: object
          properties:
            totalValue:
              type: number
              description: 총 자산 가치
            items:
              type: array
              items:
                required:
                - currentPrice
                - currentQuantity
                - currentWeight
                - diffWeight
                - expectedTradeAmount
                - recommendedQuantity
                - symbol
                - targetWeight
                type: object
                properties:
                  symbol:
                    type: string
                    description: 종목 심볼
                  currentWeight:
                    type: number
                    description: 현재 비중 (%)
                  diffWeight:
                    type: number
                    description: 비중 차이 (%p)
                  currentPrice:
                    type: number
                    description: 현재가
                  targetWeight:
                    type: number
                    description: 목표 비중 (%)
                  expectedTradeAmount:
                    type: number
                    description: 예상 매매 금액
                  recommendedQuantity:
                    type: number
                    description: 권장 보유 수량
                  currentQuantity:
                    type: number
                    description: 현재 보유 수량
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 결과 메시지
        errors:
          type: array
          description: 상세 필드 에러 목록 (성공 시 빈 리스트)
          nullable: true
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        timestamp:
          type: string
          description: 응답 시간
        status:
          type: number
          description: HTTP 상태 코드